Feat: 공통 css 분리
