+++ package.json
@@ -9,8 +9,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@tanstack/react-query": "^4.24.4",
-    "@tanstack/react-query-devtools": "^4.24.4",
     "@types/react-router-dom": "^5.3.3",
     "axios": "^1.3.2",
     "react": "^18.2.0",
