+++ src/index.tsx
@@ -1,6 +1,5 @@
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import App from './App';
 import Error from './components/Error';
@@ -15,8 +14,6 @@ import ALL from './pages/all/allPage';
 import SearchResults from './pages/search/SearchResultsPage';
 import Survey from './pages/survey/Survey';
 import Detail from './pages/detail/detailPage';
-
-const queryClient = new QueryClient();
 
 const router = createBrowserRouter([
   {
@@ -34,14 +31,14 @@ const router = createBrowserRouter([
       { path: '/survey', element: <Survey /> },
       { path: '/all', element: <ALL /> },
       { path: '/detail/:category/:id', element: <Detail /> },
+      { path: '/detail/:category/:id', element: <Detail /> },
     ],
   },
 ]);
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
-  <QueryClientProvider client={queryClient}>
+  <>
     <RouterProvider router={router} />
-    <ReactQueryDevtools initialIsOpen={true} />
-  </QueryClientProvider>
+  </>
 );
